bug fix: nao excluindo do banco de dados ao finalizar um pedido

+++ src/com/ray/model/listeners/SessionOutListener.java
@@ -8,6 +8,7 @@ import javax.servlet.http.HttpSessionListener;
 
 import com.ray.model.entities.Caneca;
 import com.ray.model.entities.Cliente;
+import com.ray.model.entities.Pedido;
 import com.ray.model.service.CanecaService;
 import com.ray.model.service.ClienteService;
 
@@ -19,15 +20,16 @@ public class SessionOutListener implements HttpSessionListener {
 	ClienteService clientService = new ClienteService();
 	CanecaService canecaService = new CanecaService();
 
+	Pedido pedido = (Pedido) se.getSession().getAttribute("order");
 	Cliente cliente = (Cliente) se.getSession().getAttribute("cliente");
 	@SuppressWarnings("unchecked")
 	List<Caneca> canecas = (List<Caneca>) se.getSession().getAttribute("canecas");
 	
-	for (Caneca c : canecas) {
-	    canecaService.deleteById(c.getId(), cliente.getId());
+	boolean hasNoOrder = pedido == null;
+	if (hasNoOrder) {
+	    canecas.forEach(x -> canecaService.deleteById(x.getId(), cliente.getId()));
+	    clientService.deleteById(cliente.getId());
 	}
-	clientService.deleteById(cliente.getId());
 	HttpSessionListener.super.sessionDestroyed(se);
     }
-
 }
